+++ lib/pragma.rb
@@ -5,9 +5,6 @@ require 'pragma/policy'
 require 'pragma/contract'
 require 'pragma/decorator'
 
-require 'will_paginate'
-require 'will_paginate/array'
-
 require 'pragma/version'
 
 require 'pragma/decorator/error'