+++ lib/pragma.rb
@@ -15,6 +15,7 @@ require 'pragma/filter/like'
 require 'pragma/filter/ilike'
 require 'pragma/filter/where'
 require 'pragma/filter/scope'
+require 'pragma/filter/boolean'
 
 require 'pragma/operation/filter'