+++ lib/pragma/macro/classes.rb
@@ -47,8 +47,9 @@ module Pragma
 
         def root_namespace(input, options)
           resource_namespace = resource_namespace(input, options)
-          return [] if resource_namespace.first == 'API'
-          resource_namespace[0..((resource_namespace.index('API') || 1) - 1)]
+          return [] if %w[API Api].include?(resource_namespace.first)
+          api_index = (resource_namespace.index('API') || resource_namespace.index('Api') || 1)
+          resource_namespace[0..(api_index - 1)]
         end
 
         def expected_model_class(input, options)