+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Implemented `Boolean` scope
+
 ## [2.3.0]
 
 ### Added
 
-- Add support for Kaminari in `Pagination` macro
+- Added support for Kaminari in `Pagination` macro
 
 ### Changed