+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+
+## [2.4.0]
 
 ### Added
 
@@ -74,7 +76,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 First Pragma 2 release.
 
-[Unreleased]: https://github.com/pragmarb/pragma/compare/v2.3.0...HEAD
+[Unreleased]: https://github.com/pragmarb/pragma/compare/v2.4.0...HEAD
+[2.4.0]: https://github.com/pragmarb/pragma/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/pragmarb/pragma/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/pragmarb/pragma/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/pragmarb/pragma/compare/v2.1.0...v2.1.1